New Universe strings & Wiki overrides
DO NOT translate the .Wiki strings. Those are used to override the search term for Wikipedia for the Universe demo when needed.

Once the demo is published, you'd be able to check which items need an override and submit those.

The .Wiki strings are excluded from being counted and marked as missing to compensate this.